+++ module-domain/src/main/java/com/devtoon/jtoon/paymentinfo/service/PaymentInfoService.java
@@ -1,9 +1,7 @@
 package com.devtoon.jtoon.paymentinfo.service;
 
-import java.io.IOException;
 import java.math.BigDecimal;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,32 +14,19 @@ import com.devtoon.jtoon.paymentinfo.entity.MemberCookie;
 import com.devtoon.jtoon.paymentinfo.entity.PaymentInfo;
 import com.devtoon.jtoon.paymentinfo.repository.MemberCookieRepository;
 import com.devtoon.jtoon.paymentinfo.repository.PaymentInfoRepository;
-import com.devtoon.jtoon.paymentinfo.request.CancelReq;
 import com.devtoon.jtoon.paymentinfo.request.PaymentReq;
-import com.siot.IamportRestClient.IamportClient;
-import com.siot.IamportRestClient.exception.IamportResponseException;
-import com.siot.IamportRestClient.request.CancelData;
 import com.siot.IamportRestClient.response.IamportResponse;
 import com.siot.IamportRestClient.response.Payment;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class PaymentInfoService {
 
-	private final IamportClient iamportClient;
 	private final PaymentInfoRepository paymentInfoRepository;
 	private final MemberCookieRepository memberCookieRepository;
-
-	public PaymentInfoService(
-		@Value("${pg.kg-inicis.rest-api-key}") String restApiKey,
-		@Value("${pg.kg-inicis.rest-api-secret}") String restSecretKey,
-		PaymentInfoRepository paymentInfoRepository,
-		MemberCookieRepository memberCookieRepository
-	) {
-		this.iamportClient = new IamportClient(restApiKey, restSecretKey);
-		this.paymentInfoRepository = paymentInfoRepository;
-		this.memberCookieRepository = memberCookieRepository;
-	}
 
 	@Transactional
 	public BigDecimal createPayment(PaymentReq paymentReq, Member member) {
@@ -54,25 +39,7 @@ public class PaymentInfoService {
 		return paymentInfo.getAmount();
 	}
 
-	public IamportResponse<Payment> cancelPayment(CancelReq cancelReq)
-		throws IamportResponseException, IOException {
-		IamportResponse<Payment> irsp = iamportClient.paymentByImpUid(cancelReq.impUid());
-		validateAmount(irsp, cancelReq.checksum());
-		CancelData cancelData = cancelReq.toCancelData(irsp);
-
-		return iamportClient.cancelPaymentByImpUid(cancelData);
-	}
-
-	public void validateIamport(PaymentReq paymentReq) throws IamportResponseException, IOException {
-		IamportResponse<Payment> irsp = iamportClient.paymentByImpUid(paymentReq.impUid());
-		CookieItem cookieItem = CookieItem.from(paymentReq.cookieItem());
-		validateAmount(irsp, cookieItem.getAmount());
-		validateAmount(irsp, paymentReq.amount());
-		validateImpUid(paymentReq);
-		validateMerchantUid(paymentReq);
-	}
-
-	private void validateAmount(IamportResponse<Payment> iamportResponse, BigDecimal amount) {
+	public void validateAmount(IamportResponse<Payment> iamportResponse, BigDecimal amount) {
 		BigDecimal realAmount = iamportResponse.getResponse().getAmount();
 
 		if (!realAmount.equals(amount)) {
@@ -80,13 +47,13 @@ public class PaymentInfoService {
 		}
 	}
 
-	private void validateMerchantUid(PaymentReq paymentReq) {
+	public void validateMerchantUid(PaymentReq paymentReq) {
 		if (paymentInfoRepository.existsByMerchantUid(paymentReq.merchantUid())) {
 			throw new DuplicatedException(ErrorStatus.PAYMENT_MERCHANT_UID_DUPLICATED);
 		}
 	}
 
-	private void validateImpUid(PaymentReq paymentReq) {
+	public void validateImpUid(PaymentReq paymentReq) {
 		if (paymentInfoRepository.existsByImpUid(paymentReq.impUid())) {
 			throw new DuplicatedException(ErrorStatus.PAYMENT_IMP_UID_DUPLICATED);
 		}
